requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
langgraph>=0.2.0
langchain-core>=0.2.0
groq>=0.9.0
pillow>=10.0.0
numpy>=1.24.0
albumentations>=1.4.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0